changes webhook available services

+++ src/webhook/webhook.controller.ts
@@ -1,29 +1,31 @@
-import { Controller, Get, Post, Body, Patch, Param, Delete , Request, HttpException, HttpStatus, Query } from '@nestjs/common';
+import { Controller, Get, Post, Body, Patch, Param, Delete, Request, HttpException, HttpStatus, Query } from '@nestjs/common';
 import { WebhookService } from './webhook.service';
-import { CreateWebhookDto } from './dto/create-webhook.dto';
+import { CreateWebhookDto, CreateWebhookDtoRequest } from './dto/create-webhook.dto';
 import { UpdateWebhookDto } from './dto/update-webhook.dto';
 import { ClientKafka, MessagePattern, Payload } from '@nestjs/microservices';
 import { Response } from 'express';
-import { OnRequest } from 'src/interface/on-request.interface'; 
+import { OnRequest } from 'src/interface/on-request.interface';
+import { FindWebhookDtoRequest } from './dto/find-webhook.dto';
+import { validateDto } from 'src/utility/validate-dto';
+import { RemoveWebhookDtoRequest } from './dto/remove.webhook.dto';
 
 const serviceFunctionMap = {
-  "webhookservice-webhook-create": "create",
+  "webhookService-webhook-create": "create",
   "webhookservice-webhook-find": "find",
   "webhookService-folder-update": "update",
+  " webhookService-webhook-remove": "remove"
 }
 
 @Controller('webhook')
 export class WebhookController {
-  constructor(private readonly webhookService: WebhookService) {}
+  constructor(private readonly webhookService: WebhookService) { }
 
-
-    
-  @Post("new")
+  @Post()
   handleHttpReq(@Request() req: any, resp: Response) {
     console.log("req.body=", req.body)
-    const request=req.body;
+    const request = req.body;
     const { topic, body, headers, method, param, query } = request;
-   
+
     if (serviceFunctionMap[topic] !== undefined)
 
       return this[serviceFunctionMap[topic]](request)
@@ -32,61 +34,41 @@ export class WebhookController {
   }
 
   @MessagePattern('webhookService-webhook-create')
-  create(@Payload() request: OnRequest) {
+  create(@Payload() request: CreateWebhookDtoRequest) {
 
-    console.log("webhook-create")
+    console.log("webhook-create", request.body)
     return this.webhookService.create(request.body);
   }
+
   
-  // @Get("get_all")
-  // @MessagePattern('webhookService-webhook-find')
-  // find(@Payload() request: OnRequest) {
-  //   return this.webhookService.find(request);
-  // }
   @MessagePattern('webhookService-webhook-find')
-  @Get('get')
-  async find(@Query('id') id?: string) {
-    try {
-      const payload: OnRequest = {
-        topic: 'webhook-fetch',
-        body: {},
-        headers: {},
-        method: 'GET',
-        param: id || '',
-        query: id ? {} : {}, 
-      };
-
-      return await this.webhookService.find(payload);
-    } catch (error) {
-      throw new HttpException(
-        {
-          statusCode: HttpStatus.INTERNAL_SERVER_ERROR,
-          status: 'Failure',
-          error: error.message || 'Internal Server Error',
-        },
-        HttpStatus.INTERNAL_SERVER_ERROR,
-      );
+ async find(@Payload() request: FindWebhookDtoRequest) {
+    const validationResult = await validateDto(FindWebhookDtoRequest, request);
+    if (validationResult) {
+      return validationResult;
     }
-  }
+   
+    return this.webhookService.find(request)
 
+  }
 
   @MessagePattern('webhookService-webhook-update')
-  @Patch('update_by_id/:id')
-  async updateById(@Param('id') id: string, @Body() updateWebhookDto: UpdateWebhookDto) {
-    return this.webhookService.update(id, updateWebhookDto); 
+  update(@Payload() request: OnRequest) {
+    return this.webhookService.update(request.body.id, request.body);
   }
 
-
-  @Delete('delete_by_id/:id')
-  async deleteById(@Param('id') id: string) {
-    return this.webhookService.remove(id); 
-  }
 
   @MessagePattern('webhookService-webhook-remove')
-  remove(@Payload() request: OnRequest) {
-    const id = request.body.id; 
-    return this.webhookService.remove(id); 
+  async remove(@Payload() request: RemoveWebhookDtoRequest) {
+
+    const validationResult = await validateDto(RemoveWebhookDtoRequest, request);
+    if (validationResult) {
+      return validationResult;
+    }
+     const id = request.body.id;
+    return this.webhookService.remove(id);
   }
+
 
 
 }
